Crate tribool

Three-valued / Three-state logic

Three-valued logic is an extension to Boolean logic with three values indicated True, False and some Indeterminate third value.

Because of the limitations of logical operator overloading in Rust, AND, OR and XOR operations are implemented with the bitwise &, | and ^ operators.

For more information and the full truth tables of this implementation, see the Wikipedia page

Re-exports

pub use Tribool::False;

pub use Tribool::Indeterminate;

pub use Tribool::True;

Structs

IndeterminateError

Error result of bool::try_from(TriBool::Indeterminate)

Enums

Tribool

Three-state Boolean logic