tribool

Enum Tribool

pub enum Tribool {
    True,
    False,
    Indeterminate,
}

Three-state Boolean logic

Variants

True

Truth value

False

False value

Indeterminate

Unknown/Indeterminate value

Implementations

impl Tribool

pub const fn boolean(value: bool) -> Tribool

Construct a new non-indeterminate Tribool from a regular boolean value

pub const fn is_true(self) -> bool

Returns true only if self is True

Example

assert!(Tribool::True.is_true())

pub const fn is_false(self) -> bool

Returns true only if self is False

Example

assert!(Tribool::False.is_false())

pub const fn is_indeterminate(self) -> bool

Returns true only if self is Indeterminate

Example

assert!(Tribool::Indeterminate.is_indeterminate())

pub const fn equals(self, rhs: Tribool) -> Tribool

Checks for equality of two Tribools, returning Indeterminate if either are indeterminate.

Example

assert!(Tribool::True.equals(Tribool::True).is_true());
assert!(Tribool::True.equals(Tribool::False).is_false());
assert!(Tribool::Indeterminate.equals(Tribool::False).is_indeterminate())

pub const fn not_equals(self, rhs: Tribool) -> Tribool

Checks for inequality of two Tribools, returning Indeterminate if either are indeterminate.

Example

assert!(Tribool::True.not_equals(Tribool::True).is_false());
assert!(Tribool::True.not_equals(Tribool::False).is_true());
assert!(Tribool::Indeterminate.not_equals(Tribool::False).is_indeterminate())

pub const fn kleene_implication(self, b: Tribool) -> Tribool

Material implication using Kleene Logic.

This is equivalent to NOT(A) OR B.

pub fn lukasiewicz_implication(self, b: Tribool) -> Tribool

Material implication using Łukasiewicz Logic

The Łukasiewicz Ł3 has the same tables for AND, OR, and NOT as the Kleene logic used elsewhere, but differs in its definition of implication in that “unknown implies unknown” is true.

For more information, see the Wikipedia page and the section on Łukasiewicz Logic

pub const fn and(self, rhs: Tribool) -> Tribool

Logical conjunction

pub const fn or(self, rhs: Tribool) -> Tribool

Logical disjunction

pub const fn xor(self, rhs: Tribool) -> Tribool

Logical exclusive disjunction

pub const fn negate(self) -> Tribool

Logical negation

pub const fn compare(self, rhs: Tribool) -> Option<Ordering>

Compares two Tribools similarly to a bool, where true is greater than false

Indeterminate values cannot be compared to order and therefore return None